moved error to shared data
