+++ api/tests/opentrons/protocol_engine/commands/flex_stacker/test_store.py
@@ -7,7 +7,6 @@ from decoy import Decoy, matchers
 
 from opentrons.drivers.flex_stacker.types import StackerAxis
 from opentrons.hardware_control.modules import FlexStacker
-from opentrons.hardware_control.modules.errors import FlexStackerStallError
 from opentrons.protocol_engine.commands.flex_stacker.common import (
     FlexStackerStallOrCollisionError,
 )
@@ -42,6 +41,7 @@ from opentrons.protocol_engine.errors import (
 )
 
 from opentrons_shared_data.labware.labware_definition import LabwareDefinition
+from opentrons_shared_data.errors.exceptions import FlexStackerStallError
 
 
 @pytest.fixture
